Don't use Serializer context
Relates:elastic/elasticsearch-net#4791

This commit updates the AttributesTableConverter and FeatureConverter
to not use the serializer context when deserializing AttributesTable.
The serializer context is not safe to use multithreaded, which can be
the case when JsonConverters from NetTopologySuite.IO.GeoJson are added
to an singleton instance of JsonSerializer not under the control
of NetTopologySuite.IO.GeoJson.